requirements.txt
asgiref==3.2.7
astroid==2.4.2
Django==3.0.7
django-environ==0.4.5
django-filter==2.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
environ==1.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
PyJWT==1.7.1
pylint==2.5.3
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
